investment_app/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
APScheduler==3.10.4

# Optional: AI model integrations
# openai==1.3.0
# anthropic==0.7.0
